+++ packages/implementations/react/src/composite/Pagination/src/index.ts
@@ -1,0 +1,5 @@
+import { Pagination } from "./Pagination";
+
+export { Pagination } from "./Pagination";
+export type { PaginationProps } from "./pagination.types";
+export default Pagination;