+++ packages/styles/src/packages/composite/pagination/rarui-pagination.types.ts
@@ -1,0 +1,7 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { paginationStyles } from "./rarui-pagination.css";
+
+export type PaginationVariants = Omit<
+  NonNullable<RecipeVariants<typeof paginationStyles.paginationItem>>,
+  "selected"
+>;